+++ lib/src/loglytics/loglytics.dart
@@ -161,10 +161,11 @@ mixin Loglytics<D extends Analytics> {
     bool fatal = false,
     bool printStack = true,
     bool addToCrashReports = true,
+    bool forceRecordError = false,
   }) {
     var _stackTrace = stackTrace ?? StackTrace.current;
     var hasError = error != null;
-    if (hasError) {
+    if (hasError || forceRecordError) {
       _crashReportsImplementation?.recordError(
         error,
         _stackTrace,